2015-04-16 Roman - !!!!:FIX:DB.PostgreSQL: Entries dublicates and other ones miss for request fieldSeek() workaround is added by "ORDER BY" to the request append 	and the entries force sort.     DateTimeDec processing fix twice: for check to type change (the row recreation) and the value for flag TCfg::ExtValTwo request. - Archive.FSArch: The messages limits into the plain format archive expanded to 100000 symbols. - UI.VCAEngine: Special IOs naming to the execution session's context is added. - FIX:UI.Vision: The primitive "Document" text global styles "font-weight" and "font-style" forcing removed and set only for the properties allowes     into the font widget's attribute, for allow working tags "<B>" and "<I>".
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2246 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/JavaLikeCalc/freefunc.cpp
@@ -395,7 +395,7 @@ int Func::regNew( bool sep, int recom )
     //Get new register
     unsigned i_rg = mRegs.size();
     if(!sep) {
-	if(recom >= 0 && recom < mRegs.size() && !mRegs[recom]->lock() && mRegs[recom]->type() == Reg::Free) i_rg = recom;
+	if(recom >= 0 && recom < (int)mRegs.size() && !mRegs[recom]->lock() && mRegs[recom]->type() == Reg::Free) i_rg = recom;
 	else for(i_rg = 0; i_rg < mRegs.size(); i_rg++)
 	    if(!mRegs[i_rg]->lock() && mRegs[i_rg]->type() == Reg::Free)
 		break;
